+++ Changelog
@@ -1,3 +1,5 @@
+==========Version 1.2==========
+1.Class separation - RGBAnalysis, RGBarray, Showft
 ==========Version 1.1==========
 1.Remove tool box switch to right click menu 
 ==========Version 1.0==========
